+++ lakehouse-traces/internal/vlstorage/vlstorage.go
@@ -89,8 +89,3 @@ func (a *adapter) DeleteActiveTasks(_ context.Context) ([]*logstorage.DeleteTask
 	return result, nil
 }
 
-func UpdatePerQueryStatsMetrics(qs *logstorage.QueryStats) {
-	if qs == nil {
-		return
-	}
-}